add AthonetServerSimulator and minor bugs fixing

+++ utils/util.py
@@ -30,4 +30,4 @@ class BandwidthConvertion():
     @classmethod
     def convert(cls, value: str = "0 Mbps", toUnit: str = "bps") -> int:
         intValue, unitValue = value.split()
-        return intValue*cls.bitTable[unitValue.lower()]/cls.bitTable[toUnit.lower()]
+        return int(int(intValue)*cls.bitTable[unitValue.lower()]/cls.bitTable[toUnit.lower()])
